requirements.txt
Flask==2.0.3
psycopg2-binary==2.9.3
tqdm==4.62.3
